extract autocomplete logic to a single file
