+++ src/Homepage/app.js
@@ -5,6 +5,7 @@ import { burgerMenu } from "./burgerMenu";
 import { togglePersona } from "./persona";
 import { cardAnimationHome } from "./cardAnimationHome";
 import { modalProduct } from "./modalProduct";
+import { autocompleteHome } from "./autocomplete";
 
 carouselGlideJS()
 GetDataForCarousel()
@@ -12,6 +13,7 @@ carouselDetailed()
 burgerMenu()
 togglePersona()
 modalProduct()
+autocompleteHome()
 
 
 var checkExist = setInterval(function () {
